Enable sandbox for chromium
This ensure that sandbox is enabled for both browsers (chromium and
firefox), an important security feature.

The reason why Playwright has sandbox disabled by default is to avoid
issues during CI, when the security features of a daily browser are not
needed. Indeed, I had to re-disable sandbox on the CI of this project.