Enable sandbox for chromium (#16)
This ensure that sandbox is enabled for both browsers (chromium and
firefox), an important security feature.

You can verify that sandbox is enabled for chromium by going to 
about:sandbox, and on firefox to about:support.

The reason why Playwright has sandbox disabled by default is to avoid
issues during CI, when the security features of a daily browser are not
needed. Indeed, I had to re-disable sandbox on the CI of this project.

Everything else should work as usual.